requirements.txt
requests
python-dotenv
flask
duffel-py